game/level: read IFF chunk-size as int32, not long
UBSan flagged the residual misaligned access at level.cc:1394 after
the HAL pool allocator fix landed. mapMem is now 8-byte aligned;
mapMem+4 is 4-aligned. Reading a `long` (8 bytes on 64-bit Linux,
LP64 ABI) from a 4-aligned address is UB.

The IFF chunk-size field on disk is 32 bits — the `long*` cast was
correct on PS1 and Win32 (where long was 4 bytes), wrong silently
on 64-bit Linux / Android / iOS. Worked anyway because the
destination `int mapStreamSize` is 4 bytes, and little-endian
truncation drops the high 4 bytes — which happen to be the next
field, not the size — so the truncated value matched the on-disk
32-bit size by coincidence.

Pinpointed via CVS: line exists verbatim in rev 1.1.1.1 of
wfsource/source/game/Attic/level.cc,v ("First source add",
2000-02-12, kts). ~26 years dormant.

Post-fix UBSan: 0 errors on snowgoons --cycles=2. Combined with
the HAL alignment fix (f10cec5), the engine is now fully clean
under -fsanitize=address,undefined for the host-GL harness.

BUGS.md entry added.

+++ wfsource/source/game/level.cc
@@ -1391,7 +1391,7 @@ Level::LoadLevelData()
 //		binistream mapStream((void*)mapMem,DiskFileCD::_SECTOR_SIZE);
 
 //		IFFChunkIter mapChunkIter(mapStream);
-		mapStreamSize = *((long*) (mapMem+4));
+		mapStreamSize = *((int32*) (mapMem+4));   // IFF chunk-size is 32-bit; was `long*` which is 8 bytes on 64-bit Linux
 		mapStreamSize += DiskFileCD::_SECTOR_SIZE - (mapStreamSize % DiskFileCD::_SECTOR_SIZE);
 		AssertMsg( mapStreamSize <= MAX_ASMP_SIZE, "ASMP chunk exceeds MAX_ASMP_SIZE; bump cap at level.cc:1296 or shrink chunk" );
 		if(mapStreamSize > DiskFileCD::_SECTOR_SIZE)
